+++ ScavengerOfPharloom/ScavengerOfPharloom.cs
@@ -8,9 +8,13 @@ namespace ScavengerOfPharloom;
 [BepInDependency("org.silksong-modding.i18n")]
 public partial class ScavengerOfPharloom : BaseUnityPlugin
 {
+    internal static ScavengerOfPharloom Instance { get; private set; }
+
     private void Awake()
     {
         // Put your initialization logic here
+        Instance = this;
+
         Harmony harmony = new Harmony(Id);
         harmony.PatchAll();
 
@@ -20,5 +24,14 @@ public partial class ScavengerOfPharloom : BaseUnityPlugin
     private void Start()
     {
         ConfigSettings.Initialize(Config);
+    }
+
+    /// <summary>
+    /// Shared logger for the mod
+    /// </summary>
+    /// <param name="message"></param>
+    internal void Log(string message)
+    {
+        Logger.LogInfo(message);
     }
 }