+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# Changelog 
+# Changelog
+## 2.0.0
+- Significantly expanded the logic so scavenged resources can be 
+found all throughout Pharloom
+- Added logic so plasmium bulbs in Wormways give lifeblood
 ## 1.0.2
 - Bug fix #2 to include languages folder
 ## 1.0.1